Implémentation de retrieveEveryPossibility.
L'implémentation n'est pas parfaite, il y a des cas normalement invalides qui
sont acceptés, et il y a des mouvements en double dans la liste retournée.

+++ strategy/include/arraylist.h
@@ -9,7 +9,7 @@ typedef struct ArrayList ArrayList;
 typedef struct {
     SMove mouvement[4] ;
     // tableau des mouvements, jusqu'à 4 possibles
-    int nombre_mouvements ;
+    size_t nombre_mouvements ;
     // nombre de mouvements viables dans cette structure ( entre 0 et 4)
 } AIListMoves;
 typedef AIListMoves TypeContent;
